Moved LED's to Matrix & Rebuilt Maze
Moved all LED's to a custom PCB that acts as an 8X8 LED matrix but does not require all LED's to be in matrix layout. This change means NO LED's need to use a single pin. All LED's are now in sequence using the MAX7219's.

Cleaned up BUZZER definition to be consistent with naming across all .h files.

Now that more pins are free due to no LED's being needed directly on a pin I moved the Load Clock and DIN pins to a more suitable position.

Moved all pin assignements for anything that was considered overlapping.

Converted Maze from MAX7219 prebuilt red only matrix to an Adafruit NeoPixel NeoMatrix 8x8. This only uses 1 data pin and only needed a few minor tweaks to work with existing code.

Added Password module complete LED logic.

Removed Venting module complete LED definition. That module should never be marked as "complete".

Moved Mins/Sec on timer back to int instead of byte due to incorrect type assignement.

+++ IRLKTANE/time.h
@@ -8,7 +8,7 @@
 
 SevenSegmentExtended timer(PIN_TIME_CLK, PIN_TIME_DIO);
 unsigned long seconds = 0;
-byte mins = BOMB_TIMER_MINUTES, sec = BOMB_TIMER_SECONDS + 1;
+int mins = BOMB_TIMER_MINUTES, sec = BOMB_TIMER_SECONDS + 1;
 
 void timeSetup() 
 {
